Complete removal of configurable matchers
